+++ lib/workers/repository/extract/index.ts
@@ -1,14 +1,11 @@
 import is from '@sindresorhus/is';
 import { getManagerConfig, mergeChildConfig } from '../../../config';
-import type {
-  ManagerConfig,
-  RenovateConfig,
-  WorkerExtractConfig,
-} from '../../../config/types';
+import type { ManagerConfig, RenovateConfig } from '../../../config/types';
 import { logger } from '../../../logger';
 import { getManagerList } from '../../../modules/manager';
 import type { PackageFile } from '../../../modules/manager/types';
 import { getFileList } from '../../../util/git';
+import type { WorkerExtractConfig } from '../../types';
 import { getMatchingFiles } from './file-match';
 import { getManagerPackageFiles } from './manager-files';
