+++ lib/workers/types.ts
@@ -1,5 +1,6 @@
 import type { Merge } from 'type-fest';
 import type {
+  CustomManager,
   GroupConfig,
   LegacyAdminConfig,
   RenovateConfig,
@@ -9,6 +10,8 @@ import type {
 import type { Release } from '../modules/datasource/types';
 import type {
   ArtifactError,
+  CustomExtractConfig,
+  ExtractConfig,
   LookupUpdate,
   PackageDependency,
   PackageFile,
@@ -122,4 +125,18 @@ export interface BranchConfig
   prNo?: number;
   stopUpdating?: boolean;
   isConflicted?: boolean;
+}
+
+export interface WorkerExtractConfig
+  extends ExtractConfig,
+    Partial<CustomExtractConfig> {
+  manager: string;
+  fileList: string[];
+  fileMatch?: string[];
+  updateInternalDeps?: boolean;
+  includePaths?: string[];
+  ignorePaths?: string[];
+  regexManagers?: CustomManager[];
+  enabledManagers?: string[];
+  enabled?: boolean;
 }
